requirements.txt
asgiref==3.8.1
boto3==1.37.7
botocore==1.37.7
diff-match-patch==20241021
Django==5.1.6
django-import-export==4.3.7
django-storages==1.14.5
djangorestframework==3.15.2
jmespath==1.0.1
mysqlclient==2.2.7
pillow==11.1.0
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
urllib3==2.3.0